[INLONG-6978][Sort] Get primary keys for PostgreSQL
