+++ inlong-sort/sort-connectors/jdbc/src/main/java/org/apache/inlong/sort/jdbc/dialect/PostgresDialect.java
@@ -17,10 +17,8 @@
 
 package org.apache.inlong.sort.jdbc.dialect;
 
-import org.apache.flink.connector.jdbc.internal.connection.SimpleJdbcConnectionProvider;
 import org.apache.flink.connector.jdbc.internal.converter.JdbcRowConverter;
 import org.apache.flink.connector.jdbc.internal.converter.PostgresRowConverter;
-import org.apache.flink.connector.jdbc.internal.options.JdbcOptions;
 import org.apache.flink.table.types.logical.LogicalTypeRoot;
 import org.apache.flink.table.types.logical.RowType;
 import org.apache.inlong.sort.jdbc.internal.JdbcMultiBatchingComm;
@@ -30,7 +28,7 @@ import org.slf4j.LoggerFactory;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
-import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
@@ -43,7 +41,7 @@ public class PostgresDialect extends AbstractJdbcDialect {
     private static final long serialVersionUID = 1L;
 
     private static final String QUERY_PK_SQL = "SELECT\n" +
-            "\tstring_agg (DISTINCT t3.attname, ',') AS pkColumn,\n" +
+            "\tstring_agg (DISTINCT t3.attname, ',') AS " + PK_COLUMN_NAME + ",\n" +
             "    \tt4.tablename AS tableName\n" +
             "FROM\n" +
             "\tpg_constraint t1\n" +
@@ -168,39 +166,10 @@ public class PostgresDialect extends AbstractJdbcDialect {
     }
 
     @Override
-    public List<String> getPkNamesFromDb(String tableIdentifier, JdbcOptions jdbcOptions) {
-        PreparedStatement st = null;
-        try {
-            JdbcOptions jdbcExecOptions = JdbcMultiBatchingComm.getExecJdbcOptions(jdbcOptions, tableIdentifier);
-            SimpleJdbcConnectionProvider tableConnectionProvider = new SimpleJdbcConnectionProvider(jdbcExecOptions);
-            Connection conn = tableConnectionProvider.getOrEstablishConnection();
-            st = conn.prepareStatement(QUERY_PK_SQL);
-            st.setString(1, JdbcMultiBatchingComm.getTbNameFromIdentifier(tableIdentifier));
-            ResultSet rs = st.executeQuery();
-            if (rs.next()) {
-                String pkColumn = rs.getString("pkColumn");
-                LOG.info("TableIdentifier:{} get pkColumn:{}", tableIdentifier, pkColumn);
-                checkAndClose(st);
-                return Arrays.asList(pkColumn.split(","));
-            } else {
-                LOG.info("TableIdentifier:{} get pkColumn: null", tableIdentifier);
-                checkAndClose(st);
-                return null;
-            }
-        } catch (Exception e) {
-            LOG.error("TableIdentifier:{} getAndSetPkNamesFromDb get err:", tableIdentifier, e);
-            checkAndClose(st);
-        }
-        return null;
-    }
-
-    private void checkAndClose(PreparedStatement st) {
-        if (null != st) {
-            try {
-                st.close();
-            } catch (Exception e) {
-                LOG.error("CheckAndClose PreparedStatement get err:", e);
-            }
-        }
+    public PreparedStatement setQuerySql(Connection conn,
+            String tableIdentifier) throws SQLException {
+        PreparedStatement st = conn.prepareStatement(QUERY_PK_SQL);
+        st.setString(1, JdbcMultiBatchingComm.getTbNameFromIdentifier(tableIdentifier));
+        return st;
     }
 }
